added close program by name and facility to open program by path and if not specified then trys to open notepad++

+++ Program.cs
@@ -5,13 +5,23 @@ class MainProgram {
         string? input = Console.ReadLine();
         if (input != null) {
             if (input.ToLower() == "o") {
-                Process.Start("C:\\Program Files\\Notepad++\\notepad++.exe");
+                Console.Write("Enter the path of the executable (if nothing provided then opening notepad++): ");
+                string? programStart = Console.ReadLine();
+                if (programStart != null) {
+                    Process.Start($"{programStart}");
+                } else {
+                    Process.Start("C:\\Program Files\\Notepad++\\notepad++.exe");
+                }
             } else if (input.ToLower() == "c") {
-                Process[] processes = Process.GetProcessesByName("notepad++");
-                foreach (Process process in processes) {
-                    process.Kill();
-                    process.WaitForExit();
-                    process.Dispose();
+                Console.Write("Enter the program name to close: ");
+                string? programName = Console.ReadLine();
+                if (programName != null) {
+                    Process[] processes = Process.GetProcessesByName($"{programName}");
+                    foreach (Process process in processes) {
+                        process.Kill();
+                        process.WaitForExit();
+                        process.Dispose();
+                    }
                 }
             }
         }
